# public/swagger.json
{"openapi":"3.1.0","servers":[{"url":"https://github.com/samchon/nestia","description":"insert your server url"}],"info":{"version":"0.0.1","title":"app-ragchat-ists","description":"A RAG chatbot for ISTS","license":{"name":"MIT"}},"paths":{"/administrator/user/edit-account":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string"},"firstNames":{"type":"string"},"lastNames":{"type":"string"},"username":{"type":"string"},"phone":{"type":"string"},"gender":{"oneOf":[{"const":"FEMENINO"},{"const":"MASCULINO"},{"const":"PREFIERO_NO_DECIR"}]},"birthDate":{"type":"string","format":"date-time"},"role":{"oneOf":[{"const":"ADMINISTRADOR"},{"const":"ANALISTA"},{"const":"ESTUDIANTE"},{"const":"DOCENTE"}]}},"required":["userId"]}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object"}}}}}}},"/administrator/user/register-account":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"firstNames":{"type":"string"},"lastNames":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"cedula":{"type":"string"},"phone":{"type":"string"},"gender":{"oneOf":[{"const":"FEMENINO"},{"const":"MASCULINO"},{"const":"PREFIERO_NO_DECIR"}]},"dateOfBirth":{"type":"string","format":"date-time"},"role":{"oneOf":[{"const":"ADMINISTRADOR"},{"const":"ANALISTA"},{"const":"ESTUDIANTE"},{"const":"DOCENTE"}]}},"required":["firstNames","lastNames","username","password","email","cedula","phone","gender","dateOfBirth","role"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o1"}}}}}}},"/user/certificate/download":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StreamableFile"}}}}}}},"/user/certificate/search":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"cedula":{"type":"string"}},"required":["cedula"]}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TryArray__object"}}}}}}},"/course/create":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"duration":{"type":"number"},"category":{"type":"string"},"level":{"oneOf":[{"const":"PRINCIPIANTE"},{"const":"INTERMEDIO"},{"const":"AVANZADO"}]},"modality":{"oneOf":[{"const":"PRESENCIAL"},{"const":"ONLINE"},{"const":"HIBRIDO"}]},"startDate":{"type":"string","format":"date-time"},"isPopular":{"type":"boolean"},"tags":{"type":"array","items":{"type":"string"}}},"required":["name","description","price","duration","category","level","modality","startDate","isPopular","tags"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o2"}}}}}}},"/course/{courseId}/delete":{"post":{"tags":[],"parameters":[{"name":"courseId","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o3"}}}}}}},"/course/edit":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"courseId":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"duration":{"type":"number"},"category":{"type":"string"},"level":{"oneOf":[{"const":"PRINCIPIANTE"},{"const":"INTERMEDIO"},{"const":"AVANZADO"}]},"modality":{"oneOf":[{"const":"PRESENCIAL"},{"const":"ONLINE"},{"const":"HIBRIDO"}]},"startDate":{"type":"string","format":"date-time"},"isPopular":{"type":"boolean"},"tags":{"type":"array","items":{"type":"string"}}},"required":["courseId"]}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o4"}}}}}}},"/course/{courseId}/recover":{"post":{"tags":[],"parameters":[{"name":"courseId","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o5"}}}}}}},"/course/search-by-filter":{"get":{"tags":[],"parameters":[{"name":"category","in":"query","schema":{"type":"string"},"required":true},{"name":"level","in":"query","schema":{"oneOf":[{"const":"PRINCIPIANTE"},{"const":"INTERMEDIO"},{"const":"AVANZADO"}]},"required":true},{"name":"modality","in":"query","schema":{"oneOf":[{"const":"PRESENCIAL"},{"const":"ONLINE"},{"const":"HIBRIDO"}]},"required":true},{"name":"page","in":"query","schema":{"type":"number"},"required":true},{"name":"perPage","in":"query","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"duration":{"type":"number"},"category":{"type":"string"},"level":{"oneOf":[{"const":"PRINCIPIANTE"},{"const":"INTERMEDIO"},{"const":"AVANZADO"}]},"modality":{"oneOf":[{"const":"PRESENCIAL"},{"const":"ONLINE"},{"const":"HIBRIDO"}]},"startDate":{"type":"string","format":"date-time"},"tags":{"oneOf":[{"type":"null"},{"type":"array","items":{"type":"string"}}]}},"required":["id","name","description","price","duration","category","level","modality","startDate","tags"]}},"totalItems":{"type":"number"},"totalPages":{"type":"number"},"perPage":{"type":"number"}},"required":["data","totalItems","totalPages","perPage"]}}}}}}},"/course/search":{"get":{"tags":[],"parameters":[{"name":"query","in":"query","schema":{"type":"string"},"required":true},{"name":"limit","in":"query","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"duration":{"type":"number"},"category":{"type":"string"},"level":{"oneOf":[{"const":"PRINCIPIANTE"},{"const":"INTERMEDIO"},{"const":"AVANZADO"}]},"modality":{"oneOf":[{"const":"PRESENCIAL"},{"const":"ONLINE"},{"const":"HIBRIDO"}]},"startDate":{"type":"string","format":"date-time"},"tags":{"oneOf":[{"type":"null"},{"type":"array","items":{"type":"string"}}]}},"required":["id","name","description","price","duration","category","level","modality","startDate","tags"]}}},"required":["data"]}}}}}}},"/course/{courseId}/view":{"get":{"tags":[],"parameters":[{"name":"courseId","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o6"}}}}}}},"/enrollment/create":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"courseId":{"type":"string"}},"required":["courseId"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o7"}}}}}}},"/enrollment/edit":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"},"hasAccess":{"type":"boolean"},"isCompleted":{"type":"boolean"}},"required":["id"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o8"}}}}}}},"/enrollment/search-by-filter":{"get":{"tags":[],"parameters":[{"name":"userId","in":"query","schema":{"type":"string"},"required":true},{"name":"courseId","in":"query","schema":{"type":"string"},"required":true},{"name":"hasAccess","in":"query","schema":{"type":"boolean"},"required":true},{"name":"isCompleted","in":"query","schema":{"type":"boolean"},"required":true},{"name":"page","in":"query","schema":{"type":"number"},"required":true},{"name":"perPage","in":"query","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"hasAccess":{"type":"boolean"},"isCompleted":{"type":"boolean"},"dateCreated":{"type":"string","format":"date-time"},"dateUpdated":{"type":"string","format":"date-time"},"user":{"oneOf":[{"type":"null"},{"type":"object","properties":{"userId":{"type":"string"},"firstNames":{"type":"string"},"lastNames":{"type":"string"},"cedula":{"type":"string"},"email":{"type":"string"}},"required":["userId","firstNames","lastNames","cedula","email"]}]},"course":{"oneOf":[{"type":"null"},{"type":"object","properties":{"courseId":{"type":"string"},"name":{"type":"string"},"category":{"type":"string"},"modality":{"oneOf":[{"const":"PRESENCIAL"},{"const":"ONLINE"},{"const":"HIBRIDO"}]},"duration":{"type":"number"}},"required":["courseId","name","category","modality","duration"]}]}},"required":["id","hasAccess","isCompleted","dateCreated","dateUpdated","user","course"]}},"totalItems":{"type":"number"},"totalPages":{"type":"number"},"perPage":{"type":"number"}},"required":["data","totalItems","totalPages","perPage"]}}}}}}},"/user/enrollment/search-by-filter":{"get":{"tags":[],"parameters":[{"name":"hasAccess","in":"query","schema":{"type":"boolean"},"required":true},{"name":"isCompleted","in":"query","schema":{"type":"boolean"},"required":true},{"name":"page","in":"query","schema":{"type":"number"},"required":true},{"name":"perPage","in":"query","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"hasAccess":{"type":"boolean"},"isCompleted":{"type":"boolean"},"dateCreated":{"type":"string","format":"date-time"},"dateUpdated":{"type":"string","format":"date-time"},"course":{"oneOf":[{"type":"null"},{"type":"object","properties":{"courseId":{"type":"string"},"name":{"type":"string"},"category":{"type":"string"},"modality":{"oneOf":[{"const":"PRESENCIAL"},{"const":"ONLINE"},{"const":"HIBRIDO"}]},"duration":{"type":"number"}},"required":["courseId","name","category","modality","duration"]}]}},"required":["id","hasAccess","isCompleted","dateCreated","dateUpdated","course"]}},"totalItems":{"type":"number"},"totalPages":{"type":"number"},"perPage":{"type":"number"}},"required":["data","totalItems","totalPages","perPage"]}}}}}}},"/facturation/create":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"legalName":{"type":"string"},"rucOrCedula":{"type":"string"},"phoneNumber":{"type":"string"},"accountingEmail":{"type":"string"},"province":{"type":"string"},"canton":{"type":"string"},"mainStreet":{"type":"string"},"addressNumber":{"type":"string"},"secondaryStreet":{"type":"string"},"isMemberOfEquinoccioNetwork":{"type":"boolean"}},"required":["legalName","rucOrCedula","phoneNumber","accountingEmail","province","canton","mainStreet","addressNumber","secondaryStreet","isMemberOfEquinoccioNetwork"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o9"}}}}}}},"/facturation/edit":{"post":{"tags":[],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o10"}}}}}}},"/facturation/{facturationId}/view":{"get":{"tags":[],"parameters":[{"name":"facturationId","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o11"}}}}}}},"/authenticate/sign-in":{"post":{"summary":"20241216 - Login with email and password","tags":["auth"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UseCaseRequest"}}},"required":true},"responses":{"200":{"description":"An encrypted token, containing a value to decode and use.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o12"}}}}}}},"/user/change-password":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"password":{"type":"string"}},"required":["password"]}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o13"}}}}}}},"/user/{userId}/delete-account":{"post":{"tags":[],"parameters":[{"name":"userId","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o14"}}}}}}},"/user/edit-account":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"firstNames":{"type":"string"},"lastNames":{"type":"string"},"username":{"type":"string"},"phone":{"type":"string"},"gender":{"oneOf":[{"const":"FEMENINO"},{"const":"MASCULINO"},{"const":"PREFIERO_NO_DECIR"}]},"dateOfBirth":{"type":"string","format":"date-time"}},"required":[]}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o15"}}}}}}},"/user/search-by-filter":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"page":{"type":"number"},"perPage":{"type":"number"},"role":{"oneOf":[{"const":"ADMINISTRADOR"},{"const":"ANALISTA"},{"const":"ESTUDIANTE"},{"const":"DOCENTE"}]}},"required":["page","perPage"]}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"role":{"oneOf":[{"const":"ADMINISTRADOR"},{"const":"ANALISTA"},{"const":"ESTUDIANTE"},{"const":"DOCENTE"}]},"cedula":{"type":"string"},"email":{"type":"string"},"firstNames":{"type":"string"},"lastNames":{"type":"string"},"phone":{"type":"string"},"gender":{"oneOf":[{"const":"FEMENINO"},{"const":"MASCULINO"},{"const":"PREFIERO_NO_DECIR"}]},"birthDate":{"type":"string","format":"date-time"},"dateCreated":{"type":"string","format":"date-time"},"dateUpdated":{"type":"string","format":"date-time"}},"required":["id","role","cedula","email","firstNames","lastNames","phone","gender","birthDate","dateCreated","dateUpdated"]}},"totalItems":{"type":"number"},"totalPages":{"type":"number"},"perPage":{"type":"number"}},"required":["data","totalItems","totalPages","perPage"]}}}}}}},"/authenticate/forgot-password":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o16"}}}}}}},"/authenticate/re-send-email-token":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o17"}}}}}}},"/user/{userId}/recover":{"post":{"tags":[],"parameters":[{"name":"userId","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o18"}}}}}}},"/authenticate/register":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"firstNames":{"type":"string"},"lastNames":{"type":"string"},"dateOfBirth":{"type":"string","format":"date-time"},"cedula":{"type":"string"},"phone":{"type":"string"},"gender":{"oneOf":[{"const":"FEMENINO"},{"const":"MASCULINO"},{"const":"PREFIERO_NO_DECIR"}]},"email":{"type":"string"},"city":{"type":"string"},"password":{"type":"string"}},"required":["firstNames","lastNames","dateOfBirth","cedula","phone","gender","email","city","password"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o19"}}}}}}},"/user/search":{"get":{"tags":[],"parameters":[{"name":"query","in":"query","schema":{"type":"string"},"required":true},{"name":"limit","in":"query","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"role":{"oneOf":[{"const":"ADMINISTRADOR"},{"const":"ANALISTA"},{"const":"ESTUDIANTE"},{"const":"DOCENTE"}]},"cedula":{"type":"string"},"email":{"type":"string"},"firstNames":{"type":"string"},"lastNames":{"type":"string"},"phone":{"type":"string"},"gender":{"oneOf":[{"const":"FEMENINO"},{"const":"MASCULINO"},{"const":"PREFIERO_NO_DECIR"}]},"birthDate":{"type":"string","format":"date-time"},"dateCreated":{"type":"string","format":"date-time"},"dateUpdated":{"type":"string","format":"date-time"}},"required":["id","role","cedula","email","firstNames","lastNames","phone","gender","birthDate","dateCreated","dateUpdated"]}}},"required":["data"]}}}}}}},"/authenticate/verify-email-token":{"post":{"tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"emailToken":{"type":"string"}},"required":["email","emailToken"]}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o20"}}}}}}},"/user/view-profile":{"get":{"tags":[],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Try__object.o21"}}}}}}}},"components":{"schemas":{"Try__object":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o1":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"StreamableFile":{"type":"object","properties":{"options":{"$ref":"#/components/schemas/StreamableFileOptions"}},"required":["options"]},"StreamableFileOptions":{"type":"object","properties":{"type":{"type":"string","title":"The value that will be used for the `Content-Type` response header","description":"The value that will be used for the `Content-Type` response header."},"disposition":{"type":"string","title":"The value that will be used for the `Content-Disposition` response header","description":"The value that will be used for the `Content-Disposition` response header."},"length":{"type":"number","title":"The value that will be used for the `Content-Length` response header","description":"The value that will be used for the `Content-Length` response header."}},"required":[],"description":"Options for `StreamableFile`"},"TryArray__object":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"course":{"type":"string"},"dateStarted":{"type":"string","format":"date-time"},"dateCompleted":{"type":"string","format":"date-time"},"duration":{"type":"number"}},"required":["id","name","course","dateStarted","dateCompleted","duration"]}},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o2":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o3":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o4":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o5":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o6":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"duration":{"type":"number"},"category":{"type":"string"},"level":{"oneOf":[{"const":"PRINCIPIANTE"},{"const":"INTERMEDIO"},{"const":"AVANZADO"}]},"modality":{"oneOf":[{"const":"PRESENCIAL"},{"const":"ONLINE"},{"const":"HIBRIDO"}]},"startDate":{"type":"string","format":"date-time"},"tags":{"oneOf":[{"type":"null"},{"type":"array","items":{"type":"string"}}]}},"required":["id","name","description","price","duration","category","level","modality","startDate","tags"]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o7":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o8":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o9":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o10":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o11":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{"id":{"type":"string"},"legalName":{"type":"string"},"rucOrCedula":{"type":"string"},"phoneNumber":{"type":"string"},"accountingEmail":{"type":"string"},"province":{"type":"string"},"canton":{"type":"string"},"mainStreet":{"type":"string"},"addressNumber":{"type":"string"},"secondaryStreet":{"type":"string"},"isMemberOfEquinoccioNetwork":{"type":"boolean"},"dateCreated":{"type":"string","format":"date-time"},"dateUpdated":{"type":"string","format":"date-time"}},"required":["id","legalName","rucOrCedula","phoneNumber","accountingEmail","province","canton","mainStreet","addressNumber","secondaryStreet","isMemberOfEquinoccioNetwork","dateCreated","dateUpdated"]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o12":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"UseCaseRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"Try__object.o13":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o14":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o15":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o16":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o17":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o18":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o19":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o20":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{},"required":[]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]},"Try__object.o21":{"type":"object","properties":{"code":{"const":"COD_OK"},"result":{"type":"object","properties":{"id":{"type":"string"},"role":{"oneOf":[{"const":"ADMINISTRADOR"},{"const":"ANALISTA"},{"const":"ESTUDIANTE"},{"const":"DOCENTE"}]},"cedula":{"type":"string"},"email":{"type":"string"},"firstNames":{"type":"string"},"lastNames":{"type":"string"},"phone":{"type":"string"},"gender":{"oneOf":[{"const":"FEMENINO"},{"const":"MASCULINO"},{"const":"PREFIERO_NO_DECIR"}]},"birthDate":{"type":"string","format":"date-time"},"dateCreated":{"type":"string","format":"date-time"},"dateUpdated":{"type":"string","format":"date-time"}},"required":["id","role","cedula","email","firstNames","lastNames","phone","gender","birthDate","dateCreated","dateUpdated"]},"info":{"type":"string"},"status":{"const":true}},"required":["code","result","info","status"]}},"securitySchemes":{"bearer":{"type":"apiKey","in":"header","name":"Authorization"}}},"tags":[{"name":"auth"}],"x-samchon-emended":true}